+++ src/lanSimulation/internals/WorkStation.java
@@ -3,7 +3,7 @@ package lanSimulation.internals;
 public class WorkStation extends NodeType {
 
 	@Override
-	public int getCode() {
+	public byte getCode() {
 		return NodeType.WORKSTATION;
 	}
